エクササイズ検索機能修正
